Update hoe hoe hoe december is here
I added the delete results method so when someone is canceling their membership they are no longer in the results. Second there was an added line of code which Chiara added that made the whole code not work so I commented it out cause I dooont really know what it did. Im sorry :(