Member v0.5
I edited the method to set the membership price very slightly.
We won't need an ENUM, it works perfectly as it is :octocat:
What we might need though is some method that limits the passive members: something that will make their membership valid only during the morning for example! :sun:

There are more suggestions at the top, in the code, check it out!